test(stark-all): simplify karma configuration in stark-* packages + showcase
ISSUES CLOSED: #1145
