+++ packages/stark-ui/karma.conf.ci.js
@@ -1,39 +1,11 @@
 const helpers = require("../stark-testing/helpers");
 
 /**
- * Load karma config from Stark
+ * Load packages karma config
  */
-const defaultKarmaCIConfig = require("../stark-testing/karma.conf.ci.js").rawKarmaConfig;
-const karmaTypescriptBundlerAlias = require("./karma.conf").karmaTypescriptBundlerAlias;
-const karmaTypescriptFiles = require("./karma.conf").karmaTypescriptFiles;
-
-// start customizing the KarmaCI configuration from stark-testing
-const starkUiSpecificConfiguration = {
-	...defaultKarmaCIConfig,
-	// change the path of the report so that Coveralls takes the right path to the source files
-	coverageIstanbulReporter: { ...defaultKarmaCIConfig.coverageIstanbulReporter, dir: helpers.root("reports/coverage/packages") },
-	// add missing files due to "@nationalbankbelgium/stark-ui" imports used in mock files of the testing sub-package
-	files: [...defaultKarmaCIConfig.files, ...karmaTypescriptFiles],
-	karmaTypescriptConfig: {
-		...defaultKarmaCIConfig.karmaTypescriptConfig,
-		bundlerOptions: {
-			...defaultKarmaCIConfig.karmaTypescriptConfig.bundlerOptions,
-			// change the module resolution for the KarmaTypescript bundler
-			resolve: {
-				alias: karmaTypescriptBundlerAlias
-			},
-			// Overwrite the karmaTypescriptConfig to pass the correct preset to karma-typescript-es6-transform
-			transforms: [
-				require("../stark-testing/node_modules/karma-typescript-angular2-transform"),
-				require("../stark-testing/node_modules/karma-typescript-es6-transform")({
-					presets: [helpers.root("../stark-testing/node_modules/babel-preset-env")] // add preset in a way that the package can find it
-				})
-			]
-		}
-	}
-};
+const rawKarmaConfig = require("../karma.conf.ci.js").rawKarmaConfig;
 
 // export the configuration function that karma expects and simply return the stark configuration
 module.exports = config => {
-	return config.set(starkUiSpecificConfiguration);
+	return config.set(rawKarmaConfig);
 };
